Removed RecoTauTag/RecoTau/data folder containing rootfiles, moved non-production code to CondFormats/PhysicsToolsObjects/test
